Move things from compute/common -> compute; add macro for opdef impl
