+++ tket-qsystem/src/extension/compute/wasm.rs
@@ -73,11 +73,9 @@ use hugr::{
 };
 use itertools::Itertools as _;
 use lazy_static::lazy_static;
-use serde::{Deserialize, Serialize};
 use smol_str::format_smolstr;
-use strum::{EnumIter, EnumString, IntoStaticStr};
 
-use super::common::{
+use super::{
     add_compute_type_defs, ComputeOp, ComputeType, CONTEXT_TYPE_NAME, FUNC_TYPE_NAME, ID_PARAM,
     INPUTS_PARAM, MODULE_TYPE_NAME, NAME_PARAM, OUTPUTS_PARAM, RESULT_TYPE_NAME,
 };
@@ -101,7 +99,7 @@ lazy_static! {
     pub static ref EXTENSION: Arc<Extension> =
         Extension::new_arc(EXTENSION_ID, EXTENSION_VERSION, |ext, ext_ref| {
         add_compute_type_defs(ext, ext_ref).unwrap();
-        WasmOpDef::load_all_ops(ext, ext_ref, ).unwrap();
+        ComputeOpDef::load_all_ops(ext, ext_ref, ).unwrap();
     });
 
     /// A [Weak] reference to the `tket.wasm` op.
@@ -113,57 +111,6 @@ lazy_static! {
         EXTENSION.to_owned(),
         PRELUDE.to_owned()
     ]);
-}
-
-#[derive(
-    Clone,
-    Copy,
-    Debug,
-    Serialize,
-    Deserialize,
-    Hash,
-    PartialEq,
-    Eq,
-    PartialOrd,
-    Ord,
-    EnumIter,
-    IntoStaticStr,
-    EnumString,
-)]
-#[allow(missing_docs, non_camel_case_types)]
-#[non_exhaustive]
-/// Simple enum of ops defined by the `tket.wasm` extension.
-pub enum WasmOpDef {
-    get_context,
-    dispose_context,
-    lookup_by_id,
-    lookup_by_name,
-    call,
-    read_result,
-}
-
-impl From<WasmType> for CustomType {
-    fn from(value: WasmType) -> Self {
-        value.custom_type(EXTENSION_ID, &EXTENSION_REF)
-    }
-}
-
-impl From<WasmType> for Type {
-    fn from(value: WasmType) -> Self {
-        value.get_type(EXTENSION_ID, &EXTENSION_REF)
-    }
-}
-
-impl TryFrom<Type> for WasmType {
-    type Error = ();
-
-    fn try_from(value: Type) -> Result<Self, Self::Error> {
-        let TypeEnum::Extension(custom_type) = value.as_type_enum() else {
-            Err(())?
-        };
-
-        custom_type.to_owned().try_into().map_err(|_| ())
-    }
 }
 
 impl TryFrom<CustomType> for WasmType {
@@ -209,350 +156,7 @@ impl TryFrom<CustomType> for WasmType {
     }
 }
 
-impl MakeOpDef for WasmOpDef {
-    fn opdef_id(&self) -> hugr::ops::OpName {
-        <&'static str>::from(self).into()
-    }
-
-    fn init_signature(&self, extension_ref: &Weak<Extension>) -> SignatureFunc {
-        let context_type = WasmType::Context.get_type(self.extension(), extension_ref);
-        let module_type = WasmType::Module.get_type(self.extension(), extension_ref);
-        match self {
-            // [usize] -> [Context]
-            Self::get_context => Signature::new(
-                usize_t(),
-                Type::from(WasmOp::get_context_return_type(
-                    self.extension(),
-                    extension_ref,
-                )),
-            )
-            .into(),
-            // [Context] -> []
-            Self::dispose_context => Signature::new(context_type, type_row![]).into(),
-            // <id: usize, inputs: TypeRow, outputs: TypeRow> [Module] -> [WasmType::Func { inputs, outputs }]
-            Self::lookup_by_id => {
-                let inputs = TypeRV::new_row_var_use(1, TypeBound::Copyable);
-                let outputs = TypeRV::new_row_var_use(2, TypeBound::Copyable);
-
-                let func_type =
-                    WasmType::func_custom_type(inputs, outputs, self.extension(), extension_ref)
-                        .into();
-                PolyFuncTypeRV::new(
-                    [
-                        ID_PARAM.to_owned(),
-                        INPUTS_PARAM.to_owned(),
-                        OUTPUTS_PARAM.to_owned(),
-                    ],
-                    Signature::new(vec![module_type], vec![func_type]),
-                )
-                .into()
-            }
-            // <name: String, inputs: TypeRow, outputs: TypeRow> [Module] -> [WasmType::Func { inputs, outputs }]
-            Self::lookup_by_name => {
-                let inputs = TypeRV::new_row_var_use(1, TypeBound::Copyable);
-                let outputs = TypeRV::new_row_var_use(2, TypeBound::Copyable);
-
-                let func_type =
-                    WasmType::func_custom_type(inputs, outputs, self.extension(), extension_ref)
-                        .into();
-                PolyFuncTypeRV::new(
-                    [
-                        NAME_PARAM.to_owned(),
-                        INPUTS_PARAM.to_owned(),
-                        OUTPUTS_PARAM.to_owned(),
-                    ],
-                    Signature::new(vec![module_type], vec![func_type]),
-                )
-                .into()
-            }
-            // <inputs: TypeRow, outputs: TypeRow> [Context, WasmType::Func { inputs, outputs }, inputs] -> [Context, future<tuple<outputs>>>]
-            Self::call => {
-                let context_type: TypeRV = context_type.into();
-                let inputs = TypeRV::new_row_var_use(0, TypeBound::Copyable);
-                let outputs = TypeRV::new_row_var_use(1, TypeBound::Copyable);
-                let func_type = Type::new_extension(WasmType::func_custom_type(
-                    inputs.clone(),
-                    outputs.clone(),
-                    self.extension(),
-                    extension_ref,
-                ));
-                let result_type = TypeRV::new_extension(WasmType::result_custom_type(
-                    outputs,
-                    self.extension(),
-                    extension_ref,
-                ));
-
-                PolyFuncTypeRV::new(
-                    [INPUTS_PARAM.to_owned(), OUTPUTS_PARAM.to_owned()],
-                    FuncValueType::new(
-                        vec![context_type.clone(), func_type.into(), inputs],
-                        vec![result_type],
-                    ),
-                )
-                .into()
-            }
-            Self::read_result => {
-                let context_type: TypeRV = context_type.into();
-                let outputs = TypeRV::new_row_var_use(0, TypeBound::Copyable);
-                let result_type = TypeRV::new_extension(WasmType::result_custom_type(
-                    outputs.clone(),
-                    self.extension(),
-                    extension_ref,
-                ));
-                PolyFuncTypeRV::new(
-                    [OUTPUTS_PARAM.to_owned()],
-                    FuncValueType::new(vec![result_type], vec![context_type, outputs]),
-                )
-                .into()
-            }
-        }
-    }
-
-    fn extension_ref(&self) -> Weak<Extension> {
-        EXTENSION_REF.to_owned()
-    }
-
-    fn from_def(op_def: &hugr::extension::OpDef) -> Result<Self, OpLoadError>
-    where
-        Self: Sized,
-    {
-        try_from_name(op_def.name().as_ref(), op_def.extension_id())
-    }
-
-    fn extension(&self) -> ExtensionId {
-        EXTENSION_ID
-    }
-}
-
-impl HasConcrete for WasmOpDef {
-    type Concrete = WasmOp;
-
-    fn instantiate(&self, type_args: &[TypeArg]) -> Result<Self::Concrete, OpLoadError> {
-        match self {
-            Self::get_context => {
-                let [] = type_args else {
-                    Err(SignatureError::from(TermTypeError::WrongNumberArgs(
-                        type_args.len(),
-                        0,
-                    )))?
-                };
-                Ok(WasmOp::GetContext)
-            }
-            Self::dispose_context => {
-                let [] = type_args else {
-                    Err(SignatureError::from(TermTypeError::WrongNumberArgs(
-                        type_args.len(),
-                        0,
-                    )))?
-                };
-                Ok(WasmOp::DisposeContext)
-            }
-            // <usize,in_row,out_row> [] -> []
-            Self::lookup_by_id => {
-                let Some([id_arg, inputs_arg, outputs_arg]): Option<[_; 3]> =
-                    type_args.to_vec().try_into().ok()
-                else {
-                    Err(SignatureError::from(TermTypeError::WrongNumberArgs(
-                        type_args.len(),
-                        3,
-                    )))?
-                };
-
-                let Some(id) = id_arg.as_nat() else {
-                    Err(SignatureError::from(TermTypeError::TypeMismatch {
-                        term: Box::new(id_arg),
-                        type_: Box::new(ID_PARAM.to_owned()),
-                    }))?
-                };
-
-                let Ok(inputs) = TypeRowRV::try_from(inputs_arg.clone()) else {
-                    Err(SignatureError::from(TermTypeError::TypeMismatch {
-                        term: Box::new(inputs_arg),
-                        type_: Box::new(INPUTS_PARAM.to_owned()),
-                    }))?
-                };
-
-                let Ok(outputs) = TypeRowRV::try_from(outputs_arg.clone()) else {
-                    Err(SignatureError::from(TermTypeError::TypeMismatch {
-                        term: Box::new(outputs_arg),
-                        type_: Box::new(OUTPUTS_PARAM.to_owned()),
-                    }))?
-                };
-                Ok(WasmOp::LookupById {
-                    id,
-                    inputs,
-                    outputs,
-                })
-            }
-            // <String,in_row,out_row> [] -> []
-            Self::lookup_by_name => {
-                let Some([name_arg, inputs_arg, outputs_arg]): Option<[_; 3]> =
-                    type_args.to_vec().try_into().ok()
-                else {
-                    Err(SignatureError::from(TermTypeError::WrongNumberArgs(
-                        type_args.len(),
-                        3,
-                    )))?
-                };
-
-                let Some(name) = name_arg.as_string() else {
-                    Err(SignatureError::from(TermTypeError::TypeMismatch {
-                        term: Box::new(name_arg),
-                        type_: Box::new(NAME_PARAM.to_owned()),
-                    }))?
-                };
-
-                let Ok(inputs) = TypeRowRV::try_from(inputs_arg.clone()) else {
-                    Err(SignatureError::from(TermTypeError::TypeMismatch {
-                        term: Box::new(inputs_arg),
-                        type_: Box::new(INPUTS_PARAM.to_owned()),
-                    }))?
-                };
-
-                let Ok(outputs) = TypeRowRV::try_from(outputs_arg.clone()) else {
-                    Err(SignatureError::from(TermTypeError::TypeMismatch {
-                        term: Box::new(outputs_arg),
-                        type_: Box::new(OUTPUTS_PARAM.to_owned()),
-                    }))?
-                };
-                Ok(WasmOp::LookupByName {
-                    name,
-                    inputs,
-                    outputs,
-                })
-            }
-            Self::call => {
-                let Some([inputs_arg, outputs_arg]): Option<[_; 2]> =
-                    type_args.to_vec().try_into().ok()
-                else {
-                    Err(SignatureError::from(TermTypeError::WrongNumberArgs(
-                        type_args.len(),
-                        2,
-                    )))?
-                };
-
-                let Ok(inputs) = TypeRowRV::try_from(inputs_arg.clone()) else {
-                    Err(SignatureError::from(TermTypeError::TypeMismatch {
-                        term: Box::new(inputs_arg),
-                        type_: Box::new(INPUTS_PARAM.to_owned()),
-                    }))?
-                };
-
-                let Ok(outputs) = TypeRowRV::try_from(outputs_arg.clone()) else {
-                    Err(SignatureError::from(TermTypeError::TypeMismatch {
-                        term: Box::new(outputs_arg),
-                        type_: Box::new(OUTPUTS_PARAM.to_owned()),
-                    }))?
-                };
-
-                Ok(WasmOp::Call {
-                    inputs: inputs.try_into()?,
-                    outputs: outputs.try_into()?,
-                })
-            }
-            Self::read_result => {
-                let Some([outputs_arg]): Option<[_; 1]> = type_args.to_vec().try_into().ok() else {
-                    Err(SignatureError::from(TermTypeError::WrongNumberArgs(
-                        type_args.len(),
-                        1,
-                    )))?
-                };
-
-                let Ok(outputs) = TypeRowRV::try_from(outputs_arg.clone()) else {
-                    Err(SignatureError::from(TermTypeError::TypeMismatch {
-                        term: Box::new(outputs_arg),
-                        type_: Box::new(OUTPUTS_PARAM.to_owned()),
-                    }))?
-                };
-                Ok(WasmOp::ReadResult {
-                    outputs: outputs.try_into()?,
-                })
-            }
-        }
-    }
-}
-
-impl TryFrom<&OpType> for WasmOpDef {
-    type Error = OpLoadError;
-
-    fn try_from(value: &OpType) -> Result<Self, Self::Error> {
-        Self::from_op(
-            value
-                .as_extension_op()
-                .ok_or(OpLoadError::NotMember(value.to_string()))?,
-        )
-    }
-}
-
-impl WasmOp {
-    fn wasm_op_def(&self) -> WasmOpDef {
-        match self {
-            Self::GetContext => WasmOpDef::get_context,
-            Self::DisposeContext => WasmOpDef::dispose_context,
-            Self::LookupById { .. } => WasmOpDef::lookup_by_id,
-            Self::LookupByName { .. } => WasmOpDef::lookup_by_name,
-            Self::Call { .. } => WasmOpDef::call,
-            Self::ReadResult { .. } => WasmOpDef::read_result,
-            Self::_Unreachable(x, _) => match *x {},
-        }
-    }
-
-    fn get_context_return_type(
-        extension_id: ExtensionId,
-        extension_ref: &Weak<Extension>,
-    ) -> SumType {
-        option_type(WasmType::Context.get_type(extension_id, extension_ref))
-    }
-}
-
-impl HasDef for WasmOp {
-    type Def = WasmOpDef;
-}
-
-impl MakeExtensionOp for WasmOp {
-    fn op_id(&self) -> OpName {
-        self.wasm_op_def().opdef_id()
-    }
-
-    fn from_extension_op(ext_op: &ExtensionOp) -> Result<Self, OpLoadError>
-    where
-        Self: Sized,
-    {
-        WasmOpDef::from_op(ext_op)?.instantiate(ext_op.args())
-    }
-
-    fn type_args(&self) -> Vec<TypeArg> {
-        match self {
-            WasmOp::GetContext => vec![],
-            WasmOp::DisposeContext => vec![],
-            WasmOp::LookupById {
-                id,
-                inputs,
-                outputs,
-            } => {
-                let inputs = TypeArg::from(inputs.clone());
-                let outputs = TypeArg::from(outputs.clone());
-                vec![TypeArg::BoundedNat(*id), inputs, outputs]
-            }
-            WasmOp::LookupByName {
-                name,
-                inputs,
-                outputs,
-            } => {
-                let inputs = TypeArg::from(inputs.clone());
-                let outputs = TypeArg::from(outputs.clone());
-                vec![name.clone().into(), inputs, outputs]
-            }
-            WasmOp::Call { inputs, outputs } => {
-                let inputs = TypeArg::from(inputs.clone());
-                let outputs = TypeArg::from(outputs.clone());
-                vec![inputs, outputs]
-            }
-            WasmOp::ReadResult { outputs } => vec![outputs.clone().into()],
-            WasmOp::_Unreachable(x, _) => match *x {},
-        }
-    }
-}
+compute_opdef!(EXTENSION_ID, WasmExtension);
 
 impl MakeRegisteredOp for WasmOp {
     fn extension_id(&self) -> ExtensionId {
